issue-#002: Documentation of the code and main documentation
The main documentation has been started and the working code documented.
At the moment, no final choices has been made regarding the type of
system will be definitely used in the future (sphinx, doxygen, a mix of
both or another solution). The choice should be taken not too late
because at the moment everything is manageable.

- Code documented using Doxygen format
- Main documentation started using Sphinx (still needs some work done)
- README file updated accordingly

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>