src/config/default.json: Add and update of the configuration file
Combination of two commits related to the configuration file with the
update of several parameters related mainly to the audio device (used by
the current version of the program) and some parameters of the FIR
filter.

src/config/default.json: add of a default configuration file

Adding of a default configuration file with few main parameters related
to the hardware setup.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>

src/config/default.json: update of the parameters

Update of the required parameters in the configuration file for the
audio device configuration.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>