src/tools.*: Update of IsFileExist and creation of GetHomeDir
Update of the function IsFileExist in order to return an error code
rather than directly the message for a better error handling (with the
idea of having a more generic system for the errors handling).

Add of a function to get the $HOME. The idea is to store a local config
file in the local home folder.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>