Drop support of python 2.6 and Pypy
As python 2.6 is officially unsupported, I don't see the need of
supporting this version. Regarding Pypy, as PyAudio doesn't support it,
I prefer to drop the support because it would require too much time keep
up to get it compatible. I added as well the support of flake8 in order
to ensure that PEP8 is supported in the code.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>